Compute Decimation recommendation for a file
Takes a file X29 or X27 from ViewDat and determines the number of
epochs and epoch interval and based on that produces a recommendation
for the decimation interval.

Was planned to read from standard in, but had a hard getting it to work
from the bash script, don’t know why.

Need to try up the usage message