Decimation now done in the source files
Previously the system decimated based on the size of file that we were
plotting, getting to throwing away 59/60 epoch for large (>4Mb) files.

This was changed for a few reasons, On some systems the browser would
not handle files more than 8Mb coming down.
It is faster to send only 1/60 of the data to the remote.

Uses less storage on the server.